refactor(TemplateBuilder): colocate styles with their consuming components
  - Create per-module styles.ts in PipeRenderer/, PipeList/, PlaceholderChip/, PipeTypeSearchMenu/ with only the styles each module uses
  - Root styles.ts reduced to editor and clearButton (TemplateQueryEditor only)
  - Remove dead styles: multiValue, multiChip, addButton
