+++ src/components/QueryEditor/TemplateBuilder/PipeRenderer/PipeFieldLoadersProvider.tsx
@@ -5,7 +5,6 @@ import { TimeRange } from '@grafana/data';
 import { VictoriaLogsDatasource } from '../../../../datasource';
 import { useFetchStreamFilters } from '../../shared/useFetchStreamFilters';
 import { useFieldFetch } from '../../shared/useFieldFetch';
-
 import { FieldLoadersProvider } from '../FieldLoadersContext';
 
 interface Props {
