+++ src/components/QueryEditor/TemplateBuilder/PipeRenderer/SegmentRenderer.tsx
@@ -3,8 +3,9 @@ import React from 'react';
 import { useStyles2 } from '@grafana/ui';
 
 import { PlaceholderChip } from '../PlaceholderChip';
-import { getStyles } from '../styles';
 import { PlaceholderSegment, Segment } from '../types';
+
+import { getStyles } from './styles';
 
 interface Props {
   segments: Segment[];
