+++ src/components/QueryEditor/TemplateBuilder/styles.ts
@@ -17,236 +17,10 @@ export const getStyles = (theme: GrafanaTheme2) => ({
     minHeight: 32,
     cursor: 'text',
   }),
-  staticText: css({
-    color: theme.colors.text.secondary,
-    whiteSpace: 'pre',
-  }),
-  pipeSeparator: css({
-    color: theme.colors.text.disabled,
-    padding: theme.spacing(0, 0.25),
-  }),
-  placeholderEmpty: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    backgroundColor: theme.colors.background.secondary,
-    borderRadius: theme.shape.radius.default,
-    padding: theme.spacing(0, 0.5),
-    color: theme.colors.text.disabled,
-    fontStyle: 'italic',
-    cursor: 'pointer',
-    minHeight: 22,
-    '&:hover': {
-      backgroundColor: theme.colors.action.hover,
-    },
-  }),
-  placeholderFilled: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    backgroundColor: theme.colors.action.selected,
-    borderRadius: theme.shape.radius.default,
-    padding: theme.spacing(0, 0.5),
-    color: theme.colors.text.primary,
-    cursor: 'pointer',
-    minHeight: 22,
-    '&:hover': {
-      backgroundColor: theme.colors.action.hover,
-    },
-  }),
-  placeholderActive: css({
-    outline: `2px solid ${theme.colors.primary.main}`,
-    outlineOffset: 1,
-  }),
-  multiValue: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    gap: theme.spacing(0.25),
-  }),
-  multiChip: css({
-    backgroundColor: theme.colors.action.selected,
-    borderRadius: theme.shape.radius.default,
-    padding: theme.spacing(0, 0.5),
-    cursor: 'pointer',
-    '&:hover': {
-      backgroundColor: theme.colors.action.hover,
-    },
-  }),
-  addButton: css({
-    border: 'none',
-    background: 'none',
-    color: theme.colors.text.secondary,
-    cursor: 'pointer',
-    padding: theme.spacing(0, 0.5),
-    '&:hover': {
-      color: theme.colors.text.primary,
-    },
-  }),
-  pipeGroup: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    gap: 2,
-  }),
-  deleteButton: css({
-    opacity: 0,
-    transition: 'opacity 0.15s',
-    '.pipe-group:hover &': {
-      opacity: 1,
-    },
-  }),
   clearButton: css({
     marginLeft: 'auto',
     opacity: 0.6,
     '&:hover': {
-      opacity: 1,
-    },
-  }),
-  chipInput: css({
-    background: 'none',
-    border: 'none',
-    outline: 'none',
-    padding: 0,
-    margin: 0,
-    fontFamily: 'inherit',
-    fontSize: 'inherit',
-    color: theme.colors.text.primary,
-    minWidth: 20,
-    width: 'auto',
-  }),
-  multiSelectedValues: css({
-    color: theme.colors.text.primary,
-    whiteSpace: 'nowrap',
-  }),
-  optionsList: css({
-    backgroundColor: theme.colors.background.primary,
-    border: `1px solid ${theme.colors.border.medium}`,
-    borderRadius: theme.shape.radius.default,
-    boxShadow: theme.shadows.z2,
-    maxHeight: 240,
-    overflowY: 'auto',
-    minWidth: 160,
-  }),
-  optionItem: css({
-    display: 'flex',
-    flexDirection: 'column',
-    padding: theme.spacing(0.5, 1),
-    cursor: 'pointer',
-  }),
-  optionItemHighlighted: css({
-    backgroundColor: theme.colors.action.hover,
-  }),
-  optionLabel: css({
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.primary,
-    fontFamily: theme.typography.fontFamilyMonospace,
-  }),
-  optionDescription: css({
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.secondary,
-    marginTop: 1,
-  }),
-  optionGroupLabel: css({
-    padding: theme.spacing(0.5, 1),
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.secondary,
-    fontWeight: theme.typography.fontWeightMedium,
-    textTransform: 'uppercase',
-    letterSpacing: '0.04em',
-    backgroundColor: theme.colors.background.secondary,
-    borderTop: `1px solid ${theme.colors.border.weak}`,
-    borderBottom: `1px solid ${theme.colors.border.weak}`,
-    '&:first-child': {
-      borderTop: 'none',
-    },
-  }),
-  pipeSearchPanel: css({
-    backgroundColor: theme.colors.background.primary,
-    border: `1px solid ${theme.colors.border.medium}`,
-    borderRadius: theme.shape.radius.default,
-    boxShadow: theme.shadows.z2,
-    width: 260,
-    display: 'flex',
-    flexDirection: 'column',
-    overflow: 'hidden',
-  }),
-  pipeSearchInputRow: css({
-    padding: theme.spacing(0.75, 1),
-    borderBottom: `1px solid ${theme.colors.border.weak}`,
-  }),
-  pipeSearchInput: css({
-    width: '100%',
-    background: 'none',
-    border: 'none',
-    outline: 'none',
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.primary,
-    '&::placeholder': {
-      color: theme.colors.text.disabled,
-    },
-  }),
-  pipeSearchList: css({
-    flex: 1,
-    minHeight: 0,
-    overflowY: 'auto',
-    padding: theme.spacing(0.5, 0),
-  }),
-  pipeSearchGroupLabel: css({
-    padding: theme.spacing(0.25, 1),
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.disabled,
-    fontWeight: theme.typography.fontWeightMedium,
-    textTransform: 'uppercase',
-    letterSpacing: '0.04em',
-    marginTop: theme.spacing(0.5),
-  }),
-  pipeSearchItem: css({
-    display: 'flex',
-    flexDirection: 'column',
-    padding: theme.spacing(0.5, 1.5),
-    cursor: 'pointer',
-  }),
-  pipeSearchItemHighlighted: css({
-    backgroundColor: theme.colors.action.hover,
-  }),
-  pipeSearchItemLabel: css({
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.primary,
-  }),
-  pipeSearchItemDesc: css({
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.secondary,
-    marginTop: 1,
-  }),
-  pipeSearchEmpty: css({
-    padding: theme.spacing(1, 1.5),
-    fontSize: theme.typography.bodySmall.fontSize,
-    color: theme.colors.text.disabled,
-  }),
-  insertSeparator: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    ['&:hover .insert-btn']: {
-      opacity: 1,
-    },
-  }),
-  insertSeparatorButton: css({
-    display: 'inline-flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: 16,
-    height: 16,
-    border: `1px solid ${theme.colors.border.medium}`,
-    borderRadius: theme.shape.radius.default,
-    background: theme.colors.background.secondary,
-    color: theme.colors.text.secondary,
-    cursor: 'pointer',
-    fontSize: 12,
-    lineHeight: 1,
-    padding: 0,
-    marginLeft: theme.spacing(0.25),
-    opacity: 0,
-    transition: 'opacity 0.15s',
-    '&:hover': {
-      background: theme.colors.action.hover,
-      color: theme.colors.text.primary,
       opacity: 1,
     },
   }),
